rename the client directory to client_dir
